+++ CHANGELOG.md
@@ -1,6 +1,11 @@
-# 0.0.0-experimental.16 - Not released yet
+# 0.0.0-experimental.16 - 15 Jul 2022
+Breaking Change:
+- Move `hook.schema` to separate plugin, [@kingworldjs/schema](https://github.com/saltyaom/kingworld-schema)
+    - To migrate, simply move all `hook.schema` to `preHandler` instead
+
 Change:
 - Rename type `ParsedRequest` to `Context`
+- Exposed `#addHandler` to `_addHandler`
 
 # 0.0.0-experimental.15 - 14 Jul 2022
 Breaking Change: